# perf(fslib): use zlib for decompression