**What's the problem this PR addresses?**

Decompression using ZipFS is slower than it has to be, this is because ZipFS uses a WebAssembly version of libzip which is limited to a single thread and does not have access to SIMD (currently behind a flag, https://v8.dev/features/simd).

Resurrects / reimplements the work started by @larixer in https://github.com/yarnpkg/berry/pull/1042

**How did you fix it?**

Used the builtin `zlib` module which uses the Node.js internal threadpool(for async reads) and SIMD to decompress the data from libzip.

**Result**

Tested on https://github.com/microsoft/TypeScript-Website/commit/6c1426fae96bc59fa0c92aa0e88047518e2a793f.

Both async and sync reads are faster and running `yarn install` with the `node-modules` linker results in ~12 seconds faster linking.

Time spent in `getFileSource`:
||time|
|--|--|
|using libzip | 20704ms
|using inflateRawSync |13222ms|
|using inflateRaw |7955ms|

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.